warehouse_mro/requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
pandas==2.2.3
openpyxl==3.1.5
XlsxWriter==3.2.0
Werkzeug==3.0.3
ultralytics==8.3.0
opencv-python==4.10.0.84
python-dotenv==1.0.1
